+++ src/LinkDotNet.StringBuilder/ValueStringBuilder.AppendFormat.cs
@@ -1,4 +1,5 @@
 using System.Diagnostics.CodeAnalysis;
+using System.Runtime.CompilerServices;
 
 namespace LinkDotNet.StringBuilder;
 
@@ -14,6 +15,7 @@ public ref partial struct ValueStringBuilder
     /// The current version does not allow for a custom format.
     /// So: <code>AppendFormat("{0:00}")</code> is not allowed and will result in an exception.
     /// </remarks>
+    [MethodImpl(MethodImplOptions.AggressiveOptimization)]
     public void AppendFormat<T>(
         [StringSyntax(StringSyntaxAttribute.CompositeFormat)] ReadOnlySpan<char> format,
         T arg)
@@ -69,6 +71,7 @@ public ref partial struct ValueStringBuilder
     /// The current version does not allow for a custom format.
     /// So: <code>AppendFormat("{0:00}")</code> is not allowed and will result in an exception.
     /// </remarks>
+    [MethodImpl(MethodImplOptions.AggressiveOptimization)]
     public void AppendFormat<T1, T2>(
         [StringSyntax(StringSyntaxAttribute.CompositeFormat)] ReadOnlySpan<char> format,
         T1 arg1,
@@ -136,6 +139,7 @@ public ref partial struct ValueStringBuilder
     /// The current version does not allow for a custom format.
     /// So: <code>AppendFormat("{0:00}")</code> is not allowed and will result in an exception.
     /// </remarks>
+    [MethodImpl(MethodImplOptions.AggressiveOptimization)]
     public void AppendFormat<T1, T2, T3>(
         [StringSyntax(StringSyntaxAttribute.CompositeFormat)] ReadOnlySpan<char> format,
         T1 arg1,
@@ -193,6 +197,7 @@ public ref partial struct ValueStringBuilder
         }
     }
 
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
     private static int GetValidArgumentIndex(ReadOnlySpan<char> placeholder, int allowedRange)
     {
 #pragma warning disable MA0011